# internal/server/openapi.yaml
openapi: 3.1.0
info:
  title: Zymo API
  version: "1"
  description: Self-hostable fermentation tracking API. All unauthorized access returns 404, not 403, to avoid leaking resource existence.
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /
    description: This instance

# All operations require auth by default. Public endpoints opt out with `security: []`.
security:
  - cookieAuth: []
  - bearerAuth: []
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: zymo_session
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    Unauthorized:
      description: Not authenticated (or session expired)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UUID:
      type: string
      format: uuid
    DateTime:
      type: string
      format: date-time
    PublicUser:
      type: object
      required: [id, username, email]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        username:
          type: string
        email:
          type: string
        display_name:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: Opaque session token. Also sent as Set-Cookie.
        user:
          $ref: '#/components/schemas/PublicUser'
    PublicProfile:
      type: object
      required: [id, username, created_at]
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        username:
          type: string
        display_name:
          type: string
        bio:
          type: string
        avatar_url:
          type: string
        created_at:
          $ref: '#/components/schemas/DateTime'
...
paths:
...
  /api/auth/register:
    post:
      tags: [auth]
      operationId: register
      summary: Register a new account
      description: |
        Behaviour depends on `INSTANCE_MODE`:
        - `open`: always allowed
        - `single_user`: allowed only while the user table is empty (bootstraps the admin)
        - `closed`: always forbidden
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                  pattern: '^[a-zA-Z0-9_-]{3,32}$'
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                display_name:
                  type: string
                  maxLength: 64
      responses:
        "201":
          description: Account created; session issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Registration disabled by instance mode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "409":
          description: Username or email already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "429":
          description: Rate limited
          headers:
            Retry-After:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Log in with username/email and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier:
                  type: string
                  description: Username or email address
                password:
                  type: string
      responses:
        "200":
          description: Authenticated; session issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "429":
          description: Rate limited
          headers:
            Retry-After:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Revoke current session
      responses:
        "200":
          description: Session revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        "401":
          $ref: '#/components/responses/Unauthorized'
  /api/auth/me:
    get:
      tags: [auth]
      operationId: getMe
      summary: Current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUser'
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # --- Users -------------------------------------------------------------------
  /api/users/{username}:
    get:
      tags: [users]
      operationId: getProfile
      summary: Profile by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicProfile'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/users/me:
    patch:
      tags: [users]
      operationId: updateProfile
      summary: Update own profile fields (COALESCE — omitted fields unchanged)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name:
                  type: string
                  maxLength: 64
                bio:
                  type: string
                  maxLength: 2048
                avatar_url:
                  type: string
                  maxLength: 512
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicProfile'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [users]
      operationId: deleteAccount
      summary: Anonymize own account (GDPR-style erasure)
      description: |
        Strips PII from the user row in place: username/email become
        derived placeholders, password and profile fields are cleared,
        and `deleted_at` is set. Sessions, push devices, notification
        prefs/notifications, and export rows + blobs are deleted.
        Recipes, comments, batches, and audit history are retained
        attached to the now-anonymized row. Refused for admin accounts —
        admin role must be handed off first. The action is logged in
        `account_deletion_requests` so a backup-restore can be
        re-anonymized via `zymo reprocess-deletions`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
      responses:
        "204":
          description: Account anonymized; session cookie cleared
        "400":
          description: Invalid JSON or account has no password set
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Password incorrect or not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "403":
          description: Admin accounts must hand off role before deletion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/users/me/password:
    post:
      tags: [users]
      operationId: changePassword
      summary: Change password; rotates all sessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
                  minLength: 8
      responses:
        "200":
          description: Password changed; new session issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Current password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/users/me/email:
    post:
      tags: [users]
      operationId: changeEmail
      summary: Request an email change (re-auth + emailed confirmation)
      description: >
        Re-authenticates with the current password and records a pending email
        change. Does NOT change the email immediately: a confirmation link is
        mailed to the new address (the change lands only when that link is
        followed) and a cancel link is mailed to the old address. Requires SMTP
        to be configured on the instance.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, email]
              properties:
                current_password:
                  type: string
                email:
                  type: string
                  format: email
      responses:
        "202":
          description: Change requested; confirmation email sent to the new address
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: pending }
                  email: { type: string, format: email }
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Current password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "409":
          description: Email already taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "502":
          description: SMTP relay error sending the confirmation email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "503":
          description: SMTP not configured on this instance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/email/confirm:
    post:
      tags: [auth]
      operationId: confirmEmailChange
      summary: Confirm a pending email change (token from the new address)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        "200":
          description: Email changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUser'
        "400":
          description: Invalid or expired link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "409":
          description: Email was claimed by another account before confirmation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/email/cancel:
    post:
      tags: [auth]
      operationId: cancelEmailChange
      summary: Cancel a pending email change (token from the old address)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        "200":
          description: Change cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: cancelled }
        "400":
          description: Invalid or expired link
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # --- Recipes -----------------------------------------------------------------
  /api/recipes:
    get:
      tags: [recipes]
      operationId: listRecipes
      summary: Recipe feed (visibility=public, newest first)
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        "200":
          description: Recipe page. `next_cursor` is null when there are no more pages.
          content:
            application/json:
              schema:
                type: object
                required: [recipes, next_cursor]
                properties:
                  recipes:
                    type: array
                    items:
                      $ref: '#/components/schemas/RecipeListItem'
                  next_cursor:
                    type: string
                    nullable: true
        "401":
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [recipes]
      operationId: createRecipe
      summary: Create a new recipe (mead, cider, or wine only in this build)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 200
                brew_type:
                  type: string
                  enum: [mead, cider, wine]
                  default: mead
                style:
                  type: string
                  maxLength: 100
                description:
                  type: string
                  maxLength: 10240
                target_og:
                  type: number
                target_fg:
                  type: number
                target_abv:
                  type: number
                batch_size_l:
                  type: number
                visibility:
                  type: string
                  enum: [public, unlisted, private]
                  default: public
                message:
                  type: string
                  description: Revision 1 message
                ingredients:
                  type: array
                  maxItems: 50
                  items:
                    $ref: '#/components/schemas/IngredientInput'
      responses:
        "201":
          description: Recipe created with revision 1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/recipes/mine:
    get:
      tags: [recipes]
      operationId: listMyRecipes
      summary: All recipes for the authenticated user (all visibilities)
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        "200":
          description: Recipe page. `next_cursor` is null when there are no more pages.
          content:
            application/json:
              schema:
                type: object
                required: [recipes, next_cursor]
                properties:
                  recipes:
                    type: array
                    items:
                      $ref: '#/components/schemas/RecipeListItem'
                  next_cursor:
                    type: string
                    nullable: true
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/recipes/{id}:
    get:
      tags: [recipes]
      operationId: getRecipe
      summary: Get recipe with live ingredients
      description: |
        Visibility rules: `public` or `unlisted` = any authed user; `private`
        = owner only. Non-owners get 404, not 403.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "200":
          description: Recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          description: Not found or not visible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags: [recipes]
      operationId: updateRecipe
      summary: Update recipe (owner only). Always creates a new revision.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 200
                style:
                  type: string
                  maxLength: 100
                description:
                  type: string
                  maxLength: 10240
                target_og:
                  type: number
                target_fg:
                  type: number
                target_abv:
                  type: number
                batch_size_l:
                  type: number
                visibility:
                  type: string
                  enum: [public, unlisted, private]
                message:
                  type: string
                ingredients:
                  type: array
                  maxItems: 50
                  items:
                    $ref: '#/components/schemas/IngredientInput'
                  description: When present, replaces all ingredients wholesale.
      responses:
        "200":
          description: Updated recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Not found or not owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [recipes]
      operationId: deleteRecipe
      summary: Delete recipe (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Not found or not owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/notifications:
    get:
      tags: [notifications]
      operationId: listNotifications
      summary: In-app notification inbox (newest first)
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        "200":
          description: Notification page. `next_cursor` is null when there are no more pages.
          content:
            application/json:
              schema:
                type: object
                required: [notifications, next_cursor]
                properties:
                  notifications:
                    type: array
                    items:
                      $ref: '#/components/schemas/Notification'
                  next_cursor:
                    type: string
                    nullable: true
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/notifications/read-all:
    post:
      tags: [notifications]
      operationId: markAllNotificationsRead
      summary: Mark all unread notifications as read
      responses:
        "204":
          description: Done
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/notifications/{id}/read:
    post:
      tags: [notifications]
      operationId: markNotificationRead
      summary: Mark a single notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        "204":
          description: Marked read
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: Notification not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...